pyproject.toml
[tool.poetry]
name = "tbtc-torch-tracker"
version = "0.1.0"
description = ""
authors = ["Michael Elsdörfer <michael@elsdoerfer.info>"]

[tool.poetry.dependencies]
python = "^3.7"
quart = "^0.13.1"
aiohttp_requests = "^0.1.3"
aiohttp = "^3.6.2"
requests = "^2.24.0"
python-dateutil = "^2.8.1"
quart-cors = "^0.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
